+++ .ci/run.sh
@@ -108,6 +108,56 @@ function run_style() {
   black --check "${SRC_ROOT}"
 }
 
+function run_docs() {
+  if [ "x${GITHUB_ACTIONS}" == "xtrue" ] && [ "x${GITHUB_REF}" != "xrefs/heads/master" ]; then
+    return
+  fi
+
+  cd "${SRC_ROOT}"
+  "${PYTHON}" -m pip install -U -e "${SRC_ROOT}[dev]"
+  "${PYTHON}" -m dffml service dev install -user
+
+  # Make master docs
+  master_docs="$(mktemp -d)"
+  TEMP_DIRS+=("${master_docs}")
+  rm -rf pages
+  ./scripts/docs.sh
+  mv pages "${master_docs}/html"
+
+  # Make last release docs
+  release_docs="$(mktemp -d)"
+  TEMP_DIRS+=("${release_docs}")
+  rm -rf pages
+  git clean -fdx
+  git checkout $(git describe --abbrev=0 --tags --match '*.*.*')
+  git clean -fdx
+  git reset --hard HEAD
+  ./scripts/docs.sh
+  mv pages "${release_docs}/html"
+
+  git clone "${SRC_ROOT}" -b gh-pages "${release_docs}/old-gh-pages-branch"
+
+  mv "${release_docs}/old-gh-pages-branch/.git" "${release_docs}/html/"
+  mv "${master_docs}/html" "${release_docs}/html/master"
+
+  cd "${release_docs}/html"
+
+  git remote set-url origin https://github.com/intel/dffml
+  git config user.name 'John Andersen'
+  git config user.email 'johnandersenpdx@gmail.com'
+  git config credential.helper store
+  git config credential.origin.username pdxjohnny
+  set +x
+  git config credential.origin.password "${GITHUB_PAGES_TOKEN}"
+  set -x
+
+  git add -A
+  git commit -sam "docs: $(date)"
+  git push
+
+  cd -
+}
+
 function cleanup_temp_dirs() {
   if [ "x${NO_RM_TEMP}" != "x" ]; then
     return
@@ -128,4 +178,6 @@ elif [ "x${WHITESPACE}" != "x" ]; then
   run_whitespace
 elif [ "x${STYLE}" != "x" ]; then
   run_style
+elif [ "x${DOCS}" != "x" ]; then
+  run_docs
 fi