+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Correctly identify when functions decorated with `op` use `self` to reference
+  the `OperationImplementationContext`.
 
 ## [0.3.1] - 2019-12-12
 ### Added