+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - GaussianProcessRegressor
     - OrthogonalMatchingPursuit
     - Lars
+    - Ridge
 - `AsyncExitStackTestCase` which instantiates and enters async and non-async
   `contextlib` exit stacks. Provides temporary file creation.
 ### Changed