+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Lars
 - `AsyncExitStackTestCase` which instantiates and enters async and non-async
   `contextlib` exit stacks. Provides temporary file creation.
+- Automatic releases to PyPi via GitHub Actions
 ### Changed
 - CLI tests and integration tests derive from `AsyncExitStackTestCase`
 ### Fixed